Moving database classes to a new package
